Add MoldUDP64 client to utilities
